requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
passlib==1.7.4
Pillow==9.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
